requirements.txt

# Core
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.6.0

# Data
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0

# Sport-specific
nba_api>=1.4.0
MLB-StatsAPI>=1.7.0
pybaseball>=2.2.0
understat>=0.1.4

# AI / Content
anthropic>=0.40.0
Pillow>=10.2.0

# Publishing
python-telegram-bot>=20.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Scheduling / async
APScheduler>=3.10.0

# Dev
pytest>=8.0.0
ruff>=0.3.0
